+++ src/PokeAByte.Domain/Interfaces/IMapperArchiveManager.cs
@@ -2,10 +2,6 @@
 
 public interface IMapperArchiveManager
 {
-    public IReadOnlyDictionary<string, IReadOnlyList<ArchivedMapperDto>> GetArchivedMappers();
-
-    public void GenerateArchivedList();
-
     public string ArchiveFile(string relativeFilename, string filepath);
 
     public void ArchiveDirectory(string directoryPath);
@@ -13,6 +9,4 @@ public interface IMapperArchiveManager
     public string BackupFile(string relativeFilename, string filepath);
 
     public void RestoreMappersFromArchive(List<ArchivedMapperDto> archivedMappers);
-
-    public void DeleteMappersFromArchive(List<ArchivedMapperDto> archivedMappers);
 }