+++ src/PokeAByte.Domain/Interfaces/IMapperFilesystemProvider.cs
@@ -1,6 +1,6 @@
 ﻿namespace PokeAByte.Domain.Interfaces;
 
-public enum MapperFilesystemTypes
+public enum MapperFilesystemTypes : byte
 {
     Official,
     Local
@@ -8,21 +8,4 @@ public enum MapperFilesystemTypes
 
 public record MapperContent(string Xml, string? ScriptPath, string? ScriptRoot);
 
-public class MapperFilesystemDTO
-{
-    public string Id { get; set; } = string.Empty;
-    public MapperFilesystemTypes Type { get; set; }
-    public string AbsolutePath { get; set; } = string.Empty;
-    public string DisplayName { get; set; } = string.Empty;
-}
-
-public interface IMapperFilesystemProvider
-{
-    IEnumerable<MapperFilesystemDTO> MapperFiles { get; }
-
-    void CacheMapperFiles();
-
-    string GetMapperRootDirectory(string absolutePath);
-    string GetRelativePath(string absolutePath);
-    Task<MapperContent> LoadContentAsync(string mapperId);
-}
+public record MapperFileData(string Id, MapperFilesystemTypes Type, string AbsolutePath, string DisplayName);